+++ src/rogallo/widgets/viewer/widget.py
@@ -24,8 +24,10 @@ from textual.app import ComposeResult
 from textual.containers import HorizontalGroup, Vertical
 from textual.events import DescendantBlur, DescendantFocus, Key
 from textual.getters import query_one
+from textual.highlight import HighlightTheme, highlight
 from textual.reactive import var
 from textual.timer import Timer
+from textual.widget import Widget
 from textual.widgets import Static
 
 ##############################################################################
@@ -43,6 +45,7 @@ from ...document import Document
 from .document_view import DocumentView
 from .gemtext import GemtextContent, GemtextLink, get_block_widget
 from .gopher import to_gemtext
+from .languages import language_from_document
 from .status import ViewerStatus
 from .title import ViewerTitle
 
@@ -166,6 +169,25 @@ class Viewer(Vertical, can_focus=False):
         if buffer:
             yield Paragraph("\n".join(buffer))
 
+    def _best_presentation_for(self, document: Document) -> list[Widget]:
+        """Get the best presentation for the document.
+
+        Args:
+            document: The document to get the best presentation for.
+
+        Returns:
+            The best presentation for the document.
+        """
+        return [
+            Static(
+                highlight(document.content, language=language, theme=HighlightTheme)
+                if not self.view_source
+                and (language := language_from_document(document))
+                else self.document.content.replace(chr(27), "\N{SYMBOL FOR ESCAPE}"),
+                markup=False,
+            )
+        ]
+
     async def _watch_document(
         self, old_document: Document, new_document: Document
     ) -> None:
@@ -185,12 +207,7 @@ class Viewer(Vertical, can_focus=False):
             await self._view.remove_children()
             self._jump_map = {}
             await self._view.mount_all(
-                [
-                    Static(
-                        self.document.content.replace(chr(27), "\N{SYMBOL FOR ESCAPE}"),
-                        markup=False,
-                    )
-                ]
+                self._best_presentation_for(self.document)
                 if not (
                     self.document.is_gemtext
                     or self.document.is_gophermap